dashboard only shows deals the user should see

+++ deal-api/src/v1/controllers/deal.controller.js
@@ -3,8 +3,17 @@ const { ObjectId } = require('mongodb');
 const moment = require('moment');
 
 const DEFAULTS = require('../defaults');
-
 const db = require('../../db-driver/client');
+const { sanitizeUser } = require('../users/sanitizeUserData');
+
+const isSuperUser = (user) => user.bank.id === '*';
+
+const dealsByOwningBank = (user) => {
+  if (isSuperUser(user)) {
+    return {};
+  }
+  return { 'details.owningBank.id': { $eq: user.bank.id } };
+};
 
 const withoutId = (obj) => {
   const cleanedObject = { ...obj };
@@ -12,10 +21,10 @@ const withoutId = (obj) => {
   return cleanedObject;
 };
 
-const findDeals = async (callback) => {
+const findDeals = async (requestingUser, callback) => {
   const collection = await db.getCollection('deals');
 
-  collection.find({})
+  collection.find(dealsByOwningBank(requestingUser))
     .sort({ updated: +1 })
     .toArray((err, result) => {
       assert.equal(err, null);
@@ -23,12 +32,12 @@ const findDeals = async (callback) => {
     });
 };
 
-const findPaginatedDeals = async (start, pagesize, callback) => {
+const findPaginatedDeals = async (requestingUser, start, pagesize, callback) => {
   const collection = await db.getCollection('deals');
 
-  const count = await collection.find({}).count();
+  const count = await collection.find(dealsByOwningBank(requestingUser)).count();
 
-  collection.find({})
+  collection.find(dealsByOwningBank(requestingUser))
     .skip(start)
     .limit(pagesize)
     .toArray((err, result) => {
@@ -43,7 +52,6 @@ const findPaginatedDeals = async (start, pagesize, callback) => {
 
 const findOneDeal = async (id, callback) => {
   const collection = await db.getCollection('deals');
-
   collection.findOne({ _id: new ObjectId(id) }, (err, result) => {
     assert.equal(err, null);
     callback(result);
@@ -61,6 +69,9 @@ exports.create = async (req, res) => {
     updated: timestamp,
   };
 
+  newDeal.details.maker = sanitizeUser(req.user);
+  newDeal.details.owningBank = req.user.bank;
+
   const response = await collection.insertOne(newDeal);
 
   const deal = response.ops[0];
@@ -68,7 +79,7 @@ exports.create = async (req, res) => {
 };
 
 exports.findAll = (req, res) => (
-  findDeals((deals) => res.status(200).send({
+  findDeals(sanitizeUser(req.user), (deals) => res.status(200).send({
     count: deals.length,
     deals,
   }))
@@ -78,27 +89,60 @@ exports.findPage = (req, res) => {
   const start = parseInt(req.params.start, 10);
   const pagesize = parseInt(req.params.pagesize, 10);
 
-  findPaginatedDeals(start, pagesize, (paginatedResults) => res.status(200).send(paginatedResults));
+  findPaginatedDeals(
+    sanitizeUser(req.user),
+    start,
+    pagesize,
+    (paginatedResults) => res.status(200).send(paginatedResults),
+  );
 };
 
-exports.findOne = (req, res) => (
-  findOneDeal(req.params.id, (deal) => res.status(200).send(deal))
-);
+exports.findOne = (req, res) => {
+  findOneDeal(req.params.id, (deal) => {
+    if (!deal) {
+      res.status(404).send();
+    } else if (isSuperUser(req.user) || (req.user.bank.id === deal.details.owningBank.id)) {
+      res.status(200).send(deal);
+    } else {
+      res.status(401).send();
+    }
+  });
+};
 
-exports.update = async (req, res) => {
-  const collection = await db.getCollection('deals');
-  await collection.updateOne(
-    { _id: { $eq: new ObjectId(req.params.id) } },
-    { $set: withoutId(req.body) }, {},
-  );
-  // TODO feels like there's a better way to achieve this...
-  const fixedDeal = { ...req.body, _id: req.params.id };
+exports.update = (req, res) => {
+  findOneDeal(req.params.id, async (deal) => {
+    if (deal) {
+      if (isSuperUser(req.user) || (req.user.bank.id === deal.details.owningBank.id)) {
+        const collection = await db.getCollection('deals');
+        await collection.updateOne(
+          { _id: { $eq: new ObjectId(req.params.id) } },
+          { $set: withoutId(req.body) }, {},
+        );
+        // TODO feels like there's a better way to achieve this...
+        const fixedDeal = { ...req.body, _id: req.params.id };
 
-  res.status(200).send(fixedDeal);
+        res.status(200).send(fixedDeal);
+      } else {
+        res.status(401).send();
+      }
+    } else {
+      res.status(404).send();
+    }
+  });
 };
 
 exports.delete = async (req, res) => {
-  const collection = await db.getCollection('deals');
-  const status = await collection.deleteOne({ _id: new ObjectId(req.params.id) });
-  res.status(200).send(status);
+  findOneDeal(req.params.id, async (deal) => {
+    if (deal) {
+      if (isSuperUser(req.user) || (req.user.bank.id === deal.details.owningBank.id)) {
+        const collection = await db.getCollection('deals');
+        const status = await collection.deleteOne({ _id: new ObjectId(req.params.id) });
+        res.status(200).send(status);
+      } else {
+        res.status(401).send();
+      }
+    } else {
+      res.status(404).send();
+    }
+  });
 };
